# Optimize spmatrix._set_many