@emcastillo 2023-10-02
Sorry, can you try to benchmark the code using this?

https://docs.cupy.dev/en/stable/user_guide/performance.html

I am interested in seeing the GPU performance. Thanks