@loganbvh 2023-10-02
@emcastillo thanks for the comments. I understand that there will still be a synchronization because of the `if`. I think that `python_scalar not in cupy.ndarray` must be doing something slow and linear in the array size (maybe copying values from device to host one by one for comparison?) to cause the performance difference I am seeing . See below

<details><summary>Test function (run on Google Colab with T4 GPU)</summary>

```python
import numpy as np
import cupy
import cupyx

def benchmark_update_only(size, nnz):

  rows = cupy.random.randint(0, size, nnz)
  cols = cupy.random.randint(0, size, nnz)
  old_vals = cupy.random.random(nnz)
  new_vals = cupy.random.random(nnz)

  mat_cupy_old = cupyx.scipy.sparse.csr_matrix((old_vals, (rows, cols)))
  mat_cupy_new = cupyx.scipy.sparse.csr_matrix((old_vals, (rows, cols)))

  def run_old():
    vals = cupy.roll(new_vals, -1)
    mat_cupy_old._set_many(rows, cols, vals)
    return mat_cupy_old.get()

  def run_new():
    # _set_many() uses mask.all()
    vals = cupy.roll(new_vals, -1)
    _set_many(mat_cupy_new, rows, cols, vals)
    return mat_cupy_new.get()

  print(f"{size = }, {nnz = }")

  print("Old _set_many:")
  print(cupyx.profiler.benchmark(run_old, (), n_repeat=100))

  print("New _set_many:")
  print(cupyx.profiler.benchmark(run_new, (), n_repeat=100))

  mat_scipy_old = run_old()
  mat_scipy_new = run_new()

  assert np.array_equal(mat_scipy_old.indices, mat_scipy_new.indices)
  assert np.array_equal(mat_scipy_old.indptr, mat_scipy_new.indptr)
  assert np.allclose(mat_scipy_old.data, mat_scipy_new.data)
  return mat_scipy_old, mat_scipy_new
```

</details>

Results:

```
size = 1000, nnz = 100
Old _set_many:
run_old             :    CPU: 8825.070 us   +/-1310.078 (min: 7724.606 / max:16055.254) us     GPU-0: 8835.807 us   +/-1311.485 (min: 7733.472 / max:16068.256) us
New _set_many:
run_new             :    CPU: 1966.075 us   +/-380.427 (min: 1650.213 / max: 3323.051) us     GPU-0: 1975.862 us   +/-381.888 (min: 1658.272 / max: 3338.944) us

size = 10000, nnz = 1000
Old _set_many:
run_old             :    CPU:77947.207 us   +/-14357.633 (min:65536.332 / max:116974.177) us     GPU-0:77964.677 us   +/-14360.006 (min:65550.400 / max:116999.359) us
New _set_many:
run_new             :    CPU: 2164.446 us   +/-354.291 (min: 1912.754 / max: 4756.252) us     GPU-0: 2174.831 us   +/-356.572 (min: 1921.440 / max: 4793.504) us

size = 100000, nnz = 1000
Old _set_many:
run_old             :    CPU:77371.187 us   +/-14551.222 (min:64965.192 / max:121929.457) us     GPU-0:77391.704 us   +/-14553.185 (min:64981.827 / max:121949.310) us
New _set_many:
run_new             :    CPU: 2201.750 us   +/-283.191 (min: 1999.019 / max: 3904.617) us     GPU-0: 2213.568 us   +/-283.910 (min: 2010.080 / max: 3918.848) us

size = 100000, nnz = 10000
Old _set_many:
run_old             :    CPU:737002.206 us   +/-101138.287 (min:659308.547 / max:1129850.700) us     GPU-0:737029.525 us   +/-101140.965 (min:659333.740 / max:1129882.568) us
New _set_many:
run_new             :    CPU: 2391.390 us   +/-299.001 (min: 2131.012 / max: 4668.803) us     GPU-0: 2403.519 us   +/-299.872 (min: 2142.144 / max: 4689.088) us
```