@loganbvh 2023-10-02
Actually, this looks to be just as fast as using `mask.all()`, maybe faster:

Note the only change from `main` is `if -1 not in offsets` --> `if -1 not in offsets.get()`. Maybe it would be worth looking into `cupy.ndarray` handling of `in` and `not in` in Python.

```python
def _set_many(self, i, j, x):
    """Sets value at each (i, j) to x
    Here (i,j) index major and minor respectively, and must not contain
    duplicate entries.
    """
    i, j, M, N = self._prepare_indices(i, j)
    x = cupy.array(x, dtype=self.dtype, copy=True, ndmin=1).ravel()

    new_sp = cupyx.scipy.sparse.csr_matrix(
        (cupy.arange(self.nnz, dtype=cupy.float32),
          self.indices, self.indptr), shape=(M, N))

    offsets = new_sp._get_arrayXarray(
        i, j, not_found_val=-1).astype(cupy.int32).ravel()

    if -1 not in offsets.get():  # This is the only line that is different from main
        # only affects existing non-zero cells
        self.data[offsets] = x
        return

    mask = offsets > -1
    self.data[offsets[mask]] = x[mask]
    # only insertions remain
    warnings.warn('Changing the sparsity structure of a '
                  '{}_matrix is expensive.'.format(self.format),
                  _base.SparseEfficiencyWarning)
    mask = ~mask
    i = i[mask]
    i[i < 0] += M
    j = j[mask]
    j[j < 0] += N
    self._insert_many(i, j, x[mask])
```

Benchmark:

```
size = 1000, nnz = 100
Old _set_many:
run_old             :    CPU: 8188.918 us   +/-1384.732 (min: 7172.135 / max:13345.911) us     GPU-0: 8199.408 us   +/-1386.683 (min: 7181.216 / max:13361.408) us
New _set_many:
run_new             :    CPU: 1409.316 us   +/-53.622 (min: 1353.613 / max: 1714.726) us     GPU-0: 1418.421 us   +/-53.911 (min: 1362.432 / max: 1723.136) us

size = 10000, nnz = 1000
Old _set_many:
run_old             :    CPU:72163.210 us   +/-12805.408 (min:61641.605 / max:107873.061) us     GPU-0:72179.694 us   +/-12807.955 (min:61653.633 / max:107892.639) us
New _set_many:
run_new             :    CPU: 1851.025 us   +/-413.465 (min: 1626.933 / max: 4501.087) us     GPU-0: 1861.673 us   +/-414.309 (min: 1635.808 / max: 4515.264) us

size = 100000, nnz = 1000
Old _set_many:
run_old             :    CPU:81232.915 us   +/-24788.292 (min:61793.035 / max:181135.444) us     GPU-0:81253.433 us   +/-24791.356 (min:61808.414 / max:181161.377) us
New _set_many:
run_new             :    CPU: 1918.286 us   +/-503.328 (min: 1617.062 / max: 5732.157) us     GPU-0: 1930.703 us   +/-504.159 (min: 1637.344 / max: 5749.920) us

size = 100000, nnz = 10000
Old _set_many:
run_old             :    CPU:689913.079 us   +/-130151.154 (min:609431.234 / max:1441613.560) us     GPU-0:689948.436 us   +/-130155.355 (min:609462.891 / max:1441665.771) us
New _set_many:
run_new             :    CPU: 1951.146 us   +/-264.610 (min: 1703.771 / max: 3905.877) us     GPU-0: 1963.304 us   +/-265.588 (min: 1714.656 / max: 3926.432) us
```